fix inconsistent scroll automation/fixing by moving both to Timeline
Before this commit, we compensated for expanding/collapsing Clusters
in ScrollFix (a contrived class component that wrapped each Cluster),
while we kicked off automatic scrolls to Items (be they instant or
smooth) from the relevant Item.

Sometimes this would break, such as when the page has focus and the
user wheels to the next Cluster, because the compensation might happen
*after* the automatic scroll, which makes the latter start from the
wrong place and immediately get cancelled.

This commit lifts all scroll management up to Timeline, converting it
to a class component (because we need getSnapshotBeforeUpdate for
scroll compensation) with a single lifecycle that allows us to put
everything in the right order.